requirements.txt
opencv-python>=4.8.0
numpy
torch>=2.0.0
ultralytics>=8.0.0
mediapipe
flask
scipy
imutils
lapx
pyyaml
requests
